Include the value of the "Range" header in the cache key
This is required for the caching handler to properly work with partial HTTP requests using the "Range" header. Without including the range in the cache key, the partial content stored first for a given URI would always be retrieved from the cache.

A test was added to ensure that two requests to the same URI with different ranges produce two entires in the cache.